# src/docs/swagger/swagger.yaml
definitions:
...
  models.Appointment:
    properties:
      id:
        example: 1
        type: integer
      location:
        example: Aula Magna
        type: string
      student:
        $ref: '#/definitions/models.Teacher'
        type: object
      time:
        type: string
    type: object
  models.Class:
    properties:
      grade:
        type: integer
      id:
        example: 1
        type: integer
      info:
        type: string
      section:
        type: string
      year:
        type: integer
    type: object
  models.Grade:
    properties:
      date:
        type: string
      grade:
        type: integer
      id:
        example: 1
        type: integer
      student:
        $ref: '#/definitions/models.Student'
        type: object
      subject:
        example: science
        type: string
      teacher:
        $ref: '#/definitions/models.Teacher'
        type: object
    type: object
...
  models.Student:
    properties:
      id:
        example: 1
        type: integer
      info:
        type: string
      mail:
        type: string
      name:
        type: string
      surname:
        type: string
    type: object
...
  representations.List:
    properties:
      data:
        items:
          type: '&{%!s(token.Pos=1100) %!s(*ast.FieldList=&{1109 [] 1110}) %!s(bool=false)}'
        type: array
      next:
        type: string
      previous:
        type: string
      self:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: These are a School management system's API .
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Back2School API
  version: "1.0"
paths:
  /appointments:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/representations.List'
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all appointments
      tags:
      - Appointments
    post:
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Appointment'
            type: object
      summary: Create appointment
      tags:
      - Appointments
  /appointments/{id}:
    get:
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Appointment'
            type: object
      summary: Get a appointment by id
      tags:
      - Appointments
    put:
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Appointment'
            type: object
      summary: Update appointment's data
      tags:
      - Appointments
  /classes:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Class'
            type: array
      summary: Get all classes
      tags:
      - Classes
    post:
      parameters:
      - description: data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Class'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Class'
            type: object
      summary: Create class
      tags:
      - Classes
  /classes/{id}:
    get:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Class'
            type: object
      summary: Get a class by id
      tags:
      - Classes
    put:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: parent
        required: true
        schema:
          $ref: '#/definitions/models.Class'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Class'
            type: object
      summary: Update Class's data
      tags:
      - Classes
  /classes/{id}/students:
    get:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
      summary: Get a student by class
      tags:
      - Classes
  /grades:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Grade'
            type: array
      summary: Get all grades
      tags:
      - Grades
    post:
      parameters:
      - description: data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Grade'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Grade'
            type: object
      summary: Create grade
      tags:
      - Grades
  /grades/{id}:
    get:
      parameters:
      - description: Grade ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Grade'
            type: object
      summary: Get a grade by id
      tags:
      - Grades
    put:
      parameters:
      - description: Grade ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Grade'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Grade'
            type: object
      summary: Update Grade
      tags:
      - Grades
  /login:
    post:
      parameters:
      - description: Add account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.Account'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Token'
            type: object
      summary: Get a login token
      tags:
      - Auth
  /notifications:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Notification'
            type: array
      summary: Get all notifications
      tags:
      - Notifications
    post:
      parameters:
      - description: data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Notification'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Notification'
            type: object
      summary: Create notification
      tags:
      - Notifications
  /notifications/{id}:
    get:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Notification'
            type: object
      summary: Get a notification by id
      tags:
      - Notifications
    put:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Notification'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Notification'
            type: object
      summary: Update notification
      tags:
      - Notifications
  /parents:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Parent'
            type: array
      summary: Get all parents
      tags:
      - Parents
    post:
      parameters:
      - description: data
        in: body
        name: parent
        required: true
        schema:
          $ref: '#/definitions/models.Parent'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Parent'
            type: object
      summary: Create parent
      tags:
      - Parents
      - Parents
  /parents/{id}:
    get:
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Parent'
            type: object
      summary: Get a parent by id
      tags:
      - Parents
    put:
      parameters:
      - description: Parent ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: parent
        required: true
        schema:
          $ref: '#/definitions/models.Parent'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Parent'
            type: object
      summary: Update parents's data
      tags:
      - Parents
  /parents/{id}/appointments:
    get:
      parameters:
      - description: Parent ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
      summary: Get appointments of the parent
      tags:
      - Parents
      - Parents
  /parents/{id}/notifications:
    get:
      parameters:
      - description: Parent ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Notification'
            type: array
      summary: Get notifications of the parent
      tags:
      - Parents
  /parents/{id}/payments:
    get:
      parameters:
      - description: Parent ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Payment'
            type: array
      summary: Get payments of the parent
      tags:
      - Parents
  /parents/{id}/students:
    get:
      parameters:
      - description: Parent ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
      summary: Get children of the parent
      tags:
      - Parents
  /payments:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Payment'
            type: array
      summary: Get all payments
      tags:
      - Payments
    post:
      parameters:
      - description: data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Payment'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Payment'
            type: object
      summary: Create payment
      tags:
      - Payments
  /payments/{id}:
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payment'
            type: object
      summary: Get a payment by id
      tags:
      - Payments
    put:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: class
        required: true
        schema:
          $ref: '#/definitions/models.Payment'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Payment'
            type: object
      summary: Update payment
      tags:
      - Payments
  /students:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
      summary: Get all students
      tags:
      - Students
    post:
      parameters:
      - description: data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Student'
            type: object
      summary: Create student
      tags:
      - Students
  /students/{id}:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
      summary: Get a lecture by class
      tags:
      - Students
    put:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/models.Student'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Student'
            type: object
      summary: Update student's data
      tags:
      - Students
  /students/{id}/grades:
    get:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Grade'
            type: array
      summary: Get grades of the student
      tags:
      - Students
  /teachers:
    get:
      parameters:
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Teacher'
            type: array
      summary: Get all teachers
      tags:
      - Teachers
    post:
      parameters:
      - description: data
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/models.Teacher'
          type: object
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Teacher'
            type: object
      summary: Create teacher
      tags:
      - Teachers
  /teachers/{id}:
    get:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Teacher'
            type: object
      summary: Get a teacher by id
      tags:
      - Teachers
    put:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: data
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/models.Teacher'
          type: object
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.Teacher'
            type: object
      summary: Update teacher's data
      tags:
      - Teachers
  /teachers/{id}/appointments:
    get:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Appointment'
            type: array
      summary: Get appointments of the teacher
      tags:
      - Teachers
  /teachers/{id}/classes:
    get:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Class'
            type: array
      summary: Get classes in which the teacher teaches
      tags:
      - Teachers
  /teachers/{id}/lectures:
    get:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TimeTable'
            type: array
      summary: Get lectures taught by the teacher
      tags:
      - Teachers
  /teachers/{id}/notifications:
    get:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: number of elements to return
        in: query
        name: limit
        type: integer
      - description: offset in the list of elements to return
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TimeTable'
            type: array
      summary: Get notifications of the teacher
      tags:
      - Teachers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"